requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-WTF==1.1.1
Werkzeug==2.3.7
SQLAlchemy==2.0.20
pymongo==4.5.0
razorpay==1.4.1
reportlab==4.0.4
python-dotenv==1.0.0
requests==2.31.0 